GCP-4041: return a long instead of an int (and I removed public from interface methods since they are not needed)
git-svn-id: https://efficio.atlassian.net/svn/GCP/trunk/IBPMiddleware@2940 4ebb2e00-0138-4608-bc6e-edf3948d4863

+++ src/main/java/org/generationcp/middleware/v2/domain/builder/ExperimentBuilder.java
@@ -30,9 +30,9 @@ public class ExperimentBuilder extends Builder {
 		super(sessionProviderForLocal, sessionProviderForCentral);
 	}
 	
-	public int count(int dataSetId) throws MiddlewareQueryException {
+	public long count(int dataSetId) throws MiddlewareQueryException {
 		if (setWorkingDatabase(dataSetId)) {
-			return (int) getExperimentProjectDao().count(dataSetId);
+			return getExperimentProjectDao().count(dataSetId);
 		}
 		return 0;
 	}
